fixed (CAST-45, CAST-46) and cleaned the Main.js
