+++ public/index.html
@@ -38,7 +38,7 @@
             <img id="logo">
         </div>
     </div>
-    <!-- <cast-media-player></cast-media-player> -->
+    <cast-media-player class="media hidden"></cast-media-player>
     <script type="module" src="./js/Main.js"></script>
     <!-- Chromecast SDK -->
     <script src="//www.gstatic.com/cast/sdk/libs/caf_receiver/v3/cast_receiver_framework.js"></script>
